+++ src/vue/forms/LimitsDocumentsUploaderForm.vue
@@ -66,7 +66,6 @@
 <script>
 import FormMixin from '@/vue/mixins/form.mixin'
 import { DOCUMENT_TYPES_TRANSLATION_IDS } from '@/js/const/document-types.const'
-import { uploadDocuments } from '@/js/helpers/upload-documents'
 import { required, documentContainer } from '@validators'
 import { Bus } from '@/js/helpers/event-bus'
 import { ErrorHandler } from '@/js/helpers/error-handler'
@@ -123,7 +122,7 @@ export default {
       this.disableForm()
       this.isRequestCreating = true
       try {
-        await uploadDocuments(Object.values(this.form.documents))
+        await base.uploadDocuments(Object.values(this.form.documents))
         await this.createLimitsUpdateRequest()
         Bus.success('limits-documents-uploader-form.documents-upload-success')
         this.$emit(EVENTS.requestUploaded)